[#72] Move base-crafting into /modules Add rune consumable type
