+++ scripts/modules/data/crafting.mjs
@@ -1,8 +1,8 @@
-import {CraftingApplication} from "../../applications/crafting-application.mjs";
-import {RecipeSheet} from "../../applications/recipe-sheet.mjs";
-import {RunesConfig} from "../../applications/runes-config.mjs";
-import {MODULE} from "../../constants.mjs";
-import {RecipeData} from "./recipe-item.mjs";
+import {CraftingApplication} from "../applications/crafting-application.mjs";
+import {RecipeSheet} from "../applications/recipe-sheet.mjs";
+import {RunesConfig} from "../applications/runes-config.mjs";
+import {MODULE} from "../constants.mjs";
+import {RecipeData} from "./models/recipe-item.mjs";
 
 /** Utility export class. */
 export class Crafting {
