+++ scripts/modules/interface.mjs
@@ -1,4 +1,4 @@
-import {Crafting} from "./data/crafting/base-crafting.mjs";
+import {Crafting} from "./data/crafting.mjs";
 import {Encounter} from "./data/encounter.mjs";
 import {Mayhem} from "./data/mayhem.mjs";
 import {ExperiencePips} from "./data/pips.mjs";
